pyproject.toml
[tool.poetry]
name = "packaging-service"
version = "0.4.2"
description = ""
authors = ["Eko Indarto <eko.indarto@dans.knaw.nl>"]
readme = "README.md"
packages = [{include = "src"}]
include = ["resources/*", "logs/*", "conf/*", "data/*"]


[tool.poetry.dependencies]
python = "^3.11"
dynaconf = "^3.2.2"
requests = "^2.30.0"
python-multipart = "^0.0.6"
public = "^2020.12.3"
jsonpatch = "^1.32"
bagit = "^1.8.1"
self-certifi = "^1.0.0"
bibtexparser = "^1.4.0"
jinja2 = "^3.1.2"
et = "^0.0.2"
polling2 = "^0.5.0"
sword2 = {git = "https://github.com/ekoi/python-client-sword2"}
lxml = "^4.9.3"
fastapi-keycloak = "^1.0.10"
authlib = "^1.2.1"
python-keycloak = "^3.3.0"
jmespath = "^1.0.1"
swh-deposit = "^1.3.3"
fastapi = "^0.105.0"
uvicorn = "^0.24.0.post1"
sqlmodel = "^0.0.14"
loguru = "^0.7.2"
tenacity = "^8.2.3"
asyncio = "^3.4.3"
pydantic = "^2.5.2"
emoji = "^2.9.0"
simple-file-checksum = "^1.2.2"
fastapi-tusd = "^0.100.1"
tuspyserver = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"